q_1546

+++ Category/baekjoon/q_1546.java
@@ -6,7 +6,7 @@ public class q_1546 {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         int n = sc.nextInt();
-        int[] scores = new int[n];
+        double[] scores = new double[n];
         double max = 0;
         for (int i = 0; i < n; i++) {
             scores[i] = sc.nextInt();
